Fix other association Ctrl+z , Ctrl+x ...
