+++ Legacy/usefull_function.py
@@ -70,3 +70,37 @@ print(f"Tempo impiegato: {end_time - start_time} secondi")
 #Ritaglio immagine, per adattarla al secondo taglio
 #x_inizio, y_inizio, larghezza, altezza = 16, 30, 455, 546
 #immagine_ritagliata = immagine[y_inizio:y_inizio + altezza, x_inizio:x_inizio + larghezza]
+
+
+# Multi process
+"""
+
+from multiprocessing import Process
+import time
+
+def worker(process_id):
+    for _ in range(5):
+        time.sleep(1)
+        print(f"Processo {process_id} in esecuzione")
+
+if __name__ == "__main__":
+    # Numero di processi da avviare
+    num_processes = 3
+
+    # Lista per mantenere riferimenti ai processi
+    processes = []
+
+    # Avvia i processi
+    for i in range(num_processes):
+        # Crea un nuovo processo con la funzione worker
+        p = Process(target=worker, args=(i,))
+        processes.append(p)
+        p.start()
+
+    # Attendi che tutti i processi completino l'esecuzione
+    for p in processes:
+        p.join()
+
+    print("Fine dello script")
+
+"""
