+++ Main.py
@@ -61,9 +61,9 @@ while(consecutive_error < 100):
     start_time = datetime.datetime.now()  #debugging tempo
 
     #screenBot.take_screenshot(870,330,490,620, label)
-#Ritaglio immagine, per adattarla al secondo taglio
-#x_inizio, y_inizio, larghezza, altezza = 16, 30, 455, 546
-#immagine_ritagliata = immagine[y_inizio:y_inizio + altezza, x_inizio:x_inizio + larghezza]
+    #Ritaglio immagine, per adattarla al secondo taglio
+    #x_inizio, y_inizio, larghezza, altezza = 16, 30, 455, 546
+    #immagine_ritagliata = immagine[y_inizio:y_inizio + altezza, x_inizio:x_inizio + larghezza]
 
 
     #Cattura screenshot
