+++ Package/solverBot.py
@@ -150,10 +150,10 @@ dizionario_movimenti = {
     'M[5][1] alto': 'alt+t',
     'M[5][2] sx': 'alt+u',
     'M[5][2] dx': 'alt+v',
-    'M[5][2] alto': 'alt+w',
-    'M[5][3] sx': 'alt+x',
-    'M[5][3] dx': 'alt+y',
-    'M[5][3] alto': 'alt+z',
+    'M[5][2] alto': 'alt+x',
+    'M[5][3] sx': 'alt+y',
+    'M[5][3] dx': 'alt+z',
+    'M[5][3] alto': 'alt+w',
     'M[5][4] sx': 'alt+1',
     'M[5][4] alto': 'alt+2',
 }
